pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autorig"
version = "1.1.0"
description = "A declarative development environment bootstrapper"
authors = [
  { name = "Dacraezy1", email = "younesaouzal18@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "jinja2>=3.0.0",
    "watchdog>=2.0.0",
    "jsonschema>=4.0.0",
    "psutil>=5.9.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort==5.12.0",
    "flake8==6.0.0",
    "mypy>=1.8.0",
    "pre-commit==3.2.0",
    "pytest>=7.0.0",
    "ruff==0.1.6",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
autorig = "autorig.cli:app"

[tool.setuptools.packages.find]
where = ["src"]